commit_20220322_23:29

+++ data/river_stations/drv_adapter_rs_data.py
@@ -451,7 +451,8 @@ class DriverData:
             folder_name_anc_list = list_folder(folder_name_anc_main)
             for folder_name_anc_step in folder_name_anc_list:
                 if os.path.exists(folder_name_anc_step):
-                    os.rmdir(folder_name_anc_step)
+                    if not os.listdir(folder_name_anc_step):
+                        os.rmdir(folder_name_anc_step)
 
     # -------------------------------------------------------------------------------------
